TA-1: more restructuring
