+++ src/app/app.module.ts
@@ -6,14 +6,14 @@ import { AngularFireDatabaseModule } from 'angularfire2/database';
 import { AngularFireStorageModule } from 'angularfire2/storage';
 import { AngularFireModule } from 'angularfire2';
 import { CityNamePipe } from './shared/city-name.pipe';
-import { ItineraryComponent } from './itinerary/itinerary/itinerary.component';
+import { ItineraryComponent } from './itinerary/itinerary.component';
 import { DayComponent } from './day/day/day.component';
 import { DayDescriptionComponent } from './day/day-description/day-description.component';
 import { DayPlacesComponent } from './day/day-places/day-places.component';
 import { DayPhotosComponent } from './day/day-photos/day-photos.component';
 import { ItineraryDayComponent } from './itinerary/itinerary-day/itinerary-day.component';
 import { DayNumberComponent } from './shared/day-number/day-number.component';
-import { ItinerarySummaryComponent } from './shared/itinerary-summary/itinerary-summary.component';
+import { ItinerarySummaryComponent } from './itinerary/itinerary-summary/itinerary-summary.component';
 
 
 @NgModule({
